% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA22204174R.R
\name{benchmarks}
\alias{benchmarks}
\title{Benchmark two approaches.}
\description{
Use R package \code{microbenchmark} to compare the performance of mixture model approach (R function \code{mixmodel_R} and Rcpp function \code{mixmodel_Rcpp}) and concave pairwise fusion approach (R function \code{concavefusion}).
}
\examples{
\dontrun{
    #parameters of the samples
    n <- 30; p <- 1; K <- 3
    sigma <- 1
    mu <- c(-2,0,3)
    beta <- 2
    #generate the samples
    set.seed(1)
    mu.sample <- sample(mu, size = n, replace = TRUE, prob = c(0.2,0.3,0.5))
    X <- rnorm(n, mean = 0, sd = sqrt(sigma))
    X <- matrix(X, nrow = n, ncol = 1)
    y <- mu.sample + X\%*\%beta + rnorm(n, mean = 0, sd =0.5)
    X1 <- cbind(1, X)
    betahat <- (solve(crossprod(X1))\%*\%crossprod(X1,y))[2,1]
    Xhat <- y-X*betahat
    #initialization or parameters
    initial_tau <- rep(1/K, K)
    initial_mu <- c(-1, 0, 1)
    initial_sigma <- rep(var(Xhat),K)
    vartheta <- 1; gamma <- 3; lambda <- 0.05; times <- 1e4
    #using the function to implement subgroup analysis
    mixmodel_R(Xhat, K, 1e5, initial_tau, initial_mu, initial_sigma)
    mixmodel_Rcpp(c(Xhat), K, 1e5, initial_tau, initial_mu, initial_sigma)
    concavefusion(X, y, times, vartheta, gamma, lambda)
    #compare the speed of these three functions
    compare.time <- microbenchmark::microbenchmark(
    mixmodel_R = mixmodel_R(Xhat, K, 1e5, initial_tau, initial_mu, initial_sigma), 
    mixmodel_Rcpp = mixmodel_Rcpp(c(Xhat), K, 1e5, initial_tau, initial_mu, initial_sigma), 
    concavefusion = concavefusion(X, y, times, vartheta, gamma, lambda))
    summary(compare.time)[,c(1,3,5,6)]
}
}
