% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA22204174R.R
\name{mixmodel_R}
\alias{mixmodel_R}
\title{A mixture model approach for cluster analysis using R}
\usage{
mixmodel_R(X, K, times = 10000, initial.tau, initial.mu, initial.sigma)
}
\arguments{
\item{X}{the random samples generated from a Gaussian mixture model, data structure: (n,p)-dimensional matrix.}

\item{K}{the number of mixture components in the population, data structure: int.}

\item{times}{the upper bound on the number of iterations, data structure: int.}

\item{initial.tau}{initial value of tau, data structure: K-dimensional vector.}

\item{initial.mu}{initial value of mu, data structure: (p,K)-dimensional matrix, or K-dimensional vector when p=1.}

\item{initial.sigma}{initial value of sigma, data structure: (p,p,K)-dimensional array, or K-dimensional vector when p=1.}
}
\value{
a list including
 iterations: the number of iterations, data structure: int.
 tau: the estimate of the proportion of each cluster, data structure: K-dimensional vector.
 mu: the estimate of expectation in each cluster, data structure: (p,K)-dimensional matrix.
 sigma: the estimate of covariance matrix in each cluster, data structure: (p,p,K)-dimensional array.
}
\description{
A mixture model approach to estimate the expectations and covariance matrices of each cluster using R
}
\examples{
\dontrun{
    #parameters of the samples
    n <- 500; p <- 2; K <- 3
    mu <- matrix(c(-2,-1,0,4,3,1), nrow = p, ncol = K)
    sigma <- matrix(0.3, nrow = p, ncol = p) + diag(0.7, p)
    #generate the random samples
    X <- matrix(0, nrow = n, ncol = p)
    set.seed(1)
    select <- sample(1:3, size = n, replace = TRUE, prob = c(1/3,1/3,1/3))
    X[select==1,] <- mvtnorm::rmvnorm(sum(select==1), mean = mu[,1], sigma = sigma)
    X[select==2,] <- mvtnorm::rmvnorm(sum(select==2), mean = mu[,2], sigma = sigma)
    X[select==3,] <- mvtnorm::rmvnorm(sum(select==3), mean = mu[,3], sigma = sigma)
    #initialization
    initial.tau <- rep(1/K, K)
    initial.mu <- matrix(c(-2,-2,0,0,2,2), nrow = p, ncol = K)
    initial.sigma <- array(rep(cov(X), K), dim = c(p,p,K))
    #cluster analysis
    mixmodel_R(X, K, times = 500, initial.tau, initial.mu, initial.sigma)
}
}
