% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mixmodel_Rcpp}
\alias{mixmodel_Rcpp}
\title{A mixture model approach for cluster analysis using Rcpp when p=1}
\usage{
mixmodel_Rcpp(X, K, times, initial_tau, initial_mu, initial_sigma)
}
\arguments{
\item{X}{the random samples generated from a Gaussian mixture model, data structure: n-dimensional vector.}

\item{K}{the number of mixture components in the population, data structure: int.}

\item{times}{the upper bound on the number of iterations, data structure: int.}

\item{initial_tau}{initial value of tau, data structure: K-dimensional vector.}

\item{initial_mu}{initial value of mu, data structure: K-dimensional vector.}

\item{initial_sigma}{initial value of sigma, data structure: K-dimensional vector.}
}
\value{
a list including
iterations: the number of iterations, data structure: int.
tau: the estimate of the proportion of each cluster, data structure: K-dimensional vector.
mu: the estimate of expectation in each cluster, data structure: K-dimensional vector.
sigma: the estimate of covariance matrix in each cluster, data structure: K-dimensional vector.
}
\description{
A mixture model approach to estimate the expectations and covariance matrices of each cluster using Rcpp when p=1
}
\examples{
\dontrun{
#parameters of the samples
n <- 300; p <- 1; K <- 3
mu <- c(-2, 0, 3)
sigma <- 1
#generate the random samples
X <- numeric(n)
set.seed(1)
select <- sample(1:3, size = n, replace = TRUE, prob = c(0.3,0.2,0.5))
X[select==1] <- rnorm(sum(select==1), mean = mu[1], sd = sqrt(sigma))
X[select==2] <- rnorm(sum(select==2), mean = mu[2], sd = sqrt(sigma))
X[select==3] <- rnorm(sum(select==3), mean = mu[3], sd = sqrt(sigma))
#initialization
initial_tau <- rep(1/K, K)
initial_mu <- c(-1, 0, 1)
initial_sigma <- rep(var(X),K)
#cluster analysis
mixmodel_Rcpp(X, K, 1e5, initial_tau, initial_mu, initial_sigma)
}
}
