\name{SA22204174-package}
\alias{SA22204174-package}
\alias{SA22204174}
\docType{package}
\title{
  Two approaches for subgroup analysis
}
\description{
  In data analysis we sometimes need to correctly identify clusters from a heterogeneous population, which is the problem named subgroup analysis. A popular method for analyzing data from a heterogeneous population is to view data as coming from a mixture of subgroups with their own sets of parameter values and then apply finite mixture model analysis. In this package, two approaches are implemented in three functions. The mixture model approach are implemented in a R function and a Rcpp function, and the concave pairwise fushion approach are implemented in a R function. A detailed description and comparison of the two approaches are also given in a vignette.
}
\author{
  SA22204174
  Maintainer: SA22204174 <wymail1202@126.com>
}
\keyword{ package }
